+++ src/main/kotlin/domain/card/Cards.kt
@@ -1,4 +1,4 @@
-package domain
+package domain.card
 
 import domain.algorithm.ScoreOptimizationAlgorithm
