+++ src/test/kotlin/blackjack/domain/PlayerTest.kt
@@ -1,9 +1,9 @@
 package blackjack.domain
 
-import domain.Card
-import domain.CardNumber
-import domain.CardShape
-import domain.Player
+import domain.card.Card
+import domain.card.CardNumber
+import domain.card.CardShape
+import domain.player.Player
 import io.kotest.core.spec.style.FreeSpec
 import io.kotest.matchers.nulls.shouldNotBeNull
 import io.kotest.matchers.shouldBe
